requirements.txt
Django==1.6.7
argparse==1.2.1
dateutils==0.6.6
django-cors-headers==0.13
django-tastypie==0.11.1
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2014.7
six==1.7.3
